+++ __test__/Header.test.tsx
@@ -5,11 +5,10 @@ import Header from '@/components/Header';
 describe('Header', () => {
   const accessToken = '';
   const refreshToken = '';
-  const setTokenData = jest.fn();
 
   it('renders kakao login button', () => {
     const { queryByText, queryByTestId } = render(
-      <Header accessToken={accessToken} refreshToken={refreshToken} setTokenData={setTokenData} />,
+      <Header accessToken={accessToken} refreshToken={refreshToken} />,
     );
 
     expect(queryByText('카카오 로그인')).not.toBeNull();
